improve ui and charts
